berberProje/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-phonenumber-field==7.2.0
gunicorn==21.2.0
phonenumbers==8.13.23
Pillow==10.1.0
python-dotenv==1.0.0
sqlparse==0.4.4
twilio==8.11.0

